replaced launch_query() to launch_scrapping() moved input_listener() from main.py to app.py
